[Feature] Allow testing for minimum versions of Git available
This patch implements a new method:

    GitWrapper::hasMinimumVersion($major, $minor, $patch)

This method makes it possible for code using the library to test
whether or not a minimum version of Git is available.  For example, we
can write...

    if ($wrapper->hasMinimumVersion(2, 5)) { ... }

...to determine if Git version 2.5.0 or later is available.  Not only
does this help library users to check for Git versions, it also allows
us to implement methods in the API which support Git commands that may
not be available on some systems depending on which version of Git the
developer has available, e.g. `git worktree`, which requires Git 2.5
or later.

Signed-off-by: Eric James Michael Ritz <ejmr@plutono.com>